+++ test/server/routes/DefaultRoute.ts
@@ -1,8 +1,9 @@
 ///////////////////////////////////////////////////////////////////////////////
 // Copyright (C) 2018 AcceleratXR, Inc. All rights reserved.
 ///////////////////////////////////////////////////////////////////////////////
-import { Route, Get, User, Init, Auth, WebSocket, Socket } from "../../../src/decorators/RouteDecorators";
+import { Route, Get, User, Auth, WebSocket, Socket, Query } from "../../../src/decorators/RouteDecorators";
 import { ApiError, Logger } from "@composer-js/core";
+import { Init } from "../../../src/decorators/ObjectDecorators";
 import * as ws from "ws";
 import { Description, Returns } from "../../../src/decorators/DocDecorators";
 import { ApiErrors, ApiErrorMessages } from "../../../src/ApiErrors";
@@ -15,12 +16,12 @@ class DefaultRoute {
     /**
      * Initializes a new instance with the specified defaults.
      */
-    constructor() { 
+    constructor() {
         // NO-OP
     }
 
     @Init
-    private async initialize() { 
+    private async initialize() {
         // NO-OP
     }
 
@@ -67,6 +68,15 @@ class DefaultRoute {
         } else {
             throw new ApiError(ApiErrors.AUTH_REQUIRED, 401, ApiErrorMessages.AUTH_REQUIRED);
         }
+    }
+
+    @WebSocket("connect-query")
+    @Description("Establishes a socket connection that responds to all messages with the query message and message `echo ${message} ${msg}` or `echo ${msg}`.")
+    protected wsConnectQuery(@Query("message") message, @Socket ws: ws, @User user?: any): void {
+        ws.on("message", (msg) => {
+            ws.send(`echo ${message ? `${message} ` : ''}${msg}`);
+        });
+        ws.send(`hello ${user && user.uid ? user.uid : "guest"}`);
     }
 }